move `init_tracing` outside of the scanner
